아이템 작성 및 생성

+++ Assets/Scripts/Entities/CharacterStats.cs
@@ -15,7 +15,7 @@ public class CharacterStats
 {
     public eStatsChangeType statsChangeType;
     [SerializeField, Range(1, 100)] public int maxHealth;
-    [SerializeField, Range(1f, 100f)] public int speed;
+    [SerializeField, Range(1f, 100f)] public float speed;
 
     // 공격 데이터
     public AttackSO attackSO;
